Separated out some legacy rendering utils as helper functions
