+++ src/glview/Renderer.h
@@ -80,11 +80,9 @@ public:
   virtual Color4f setColor(ColorMode colormode, const float color[4], const shaderinfo_t *shaderinfo = nullptr) const;
   virtual void setColorScheme(const ColorScheme& cs);
 
-  [[nodiscard]] virtual csgmode_e get_csgmode(const bool highlight_mode, const bool background_mode, const OpenSCADOperator type = OpenSCADOperator::UNION) const;
-  virtual void render_surface(const PolySet& geom, csgmode_e csgmode, const Transform3d& m, const shaderinfo_t *shaderinfo = nullptr) const;
-  virtual void render_edges(const PolySet& geom, csgmode_e csgmode) const;
   virtual std::vector<SelectedObject> findModelObject(Vector3d near_pt, Vector3d far_pt, int mouse_x, int mouse_y, double tolerance);
 
+  [[nodiscard]] static csgmode_e get_csgmode(const bool highlight_mode, const bool background_mode, const OpenSCADOperator type = OpenSCADOperator::UNION);
 protected:
   std::map<ColorMode, Color4f> colormap;
   const ColorScheme *colorscheme{nullptr};
